+++ client-app/src/desktop/AppModel.ts
@@ -12,6 +12,7 @@ import {Icon} from '@xh/hoist/icon';
 import {runInAction} from '@xh/hoist/mobx';
 import {isEmpty} from 'lodash';
 import {BaseAppModel} from '../BaseAppModel';
+import {DocService} from '../core/svc/DocService';
 import {GitHubService} from '../core/svc/GitHubService';
 import {PortfolioService} from '../core/svc/PortfolioService';
 import {
@@ -21,8 +22,9 @@ import {
     simpleTreeMapPanel,
     splitTreeMapPanel
 } from './tabs/charts';
+import {docsTab} from './tabs/docs/DocsTab';
 import {examplesTab} from './tabs/examples/ExamplesTab';
-import {formPanel, inputsPanel, toolbarFormPanel} from './tabs/forms';
+import {formPanel, inputsPanel, pickerPanel, selectPanel, toolbarFormPanel} from './tabs/forms';
 import {
     agGridView,
     columnFilteringPanel,
@@ -85,7 +87,7 @@ export class AppModel extends BaseAppModel {
 
     override async initAsync() {
         await super.initAsync();
-        await XH.installServicesAsync(GitHubService, PortfolioService);
+        await XH.installServicesAsync(DocService, GitHubService, PortfolioService);
 
         // Demo app-specific handling of EnvironmentService.serverVersion observable.
         this.addReaction({
@@ -195,6 +197,8 @@ export class AppModel extends BaseAppModel {
                         children: [
                             {name: 'form', path: '/form'},
                             {name: 'inputs', path: '/inputs'},
+                            {name: 'select', path: '/select'},
+                            {name: 'picker', path: '/picker'},
                             {name: 'toolbarForm', path: '/toolbarForm'}
                         ]
                     },
@@ -240,6 +244,11 @@ export class AppModel extends BaseAppModel {
                             },
                             {name: 'timestamp', path: '/timestamp'}
                         ]
+                    },
+                    {
+                        name: 'docs',
+                        path: '/docs',
+                        children: [{name: 'docId', path: '/:docId'}]
                     },
                     {
                         name: 'examples',
@@ -361,6 +370,8 @@ export class AppModel extends BaseAppModel {
                     tabs: [
                         {id: 'form', title: 'FormModel', content: formPanel},
                         {id: 'inputs', title: 'Hoist Inputs', content: inputsPanel},
+                        {id: 'select', title: 'Select', content: selectPanel},
+                        {id: 'picker', title: 'Picker', content: pickerPanel},
                         {id: 'toolbarForm', title: 'Toolbar Forms', content: toolbarFormPanel}
                     ]
                 }
@@ -415,6 +426,7 @@ export class AppModel extends BaseAppModel {
                     ]
                 }
             },
+            {id: 'docs', icon: Icon.book(), content: docsTab},
             {id: 'examples', icon: Icon.books(), content: examplesTab}
         ];
         return new TabContainerModel({