+++ client-app/src/desktop/tabs/grids/ColumnGroupsGridPanel.ts
@@ -1,5 +1,7 @@
-import {hoistCmp} from '@xh/hoist/core';
-import {a, em, p} from '@xh/hoist/cmp/layout';
+import {elementFactory, hoistCmp} from '@xh/hoist/core';
+import {a, p} from '@xh/hoist/cmp/layout';
+
+const em = elementFactory('em');
 import {panel} from '@xh/hoist/desktop/cmp/panel';
 import {Icon} from '@xh/hoist/icon';
 import {sampleColumnGroupsGrid, wrapper} from '../../common';
@@ -17,12 +19,12 @@ export const columnGroupsGridPanel = hoistCmp.factory(() =>
                 '.'
             ),
             p(
-                'Note that column group configurations must be provided either a headerName or groupId property, which must be unique within the GridModel. Column groups in Hoist React are also \'sealed\', meaning that columns may be reordered ',
+                "Note that column group configurations must be provided either a headerName or groupId property, which must be unique within the GridModel. Column groups in Hoist React are also 'sealed', meaning that columns may be reordered ",
                 em('within'),
                 ' the group in which they are defined but not broken out from them.'
             ),
             p(
-                'This grid also demonstrates persistent grid state using the browser\'s local storage.'
+                "This grid also demonstrates persistent grid state using the browser's local storage."
             )
         ],
         item: panel({